% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident locations on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number coerced to integer correspnding to an USA state.}

\item{year}{A number representing the data for a specific year. This cannot
be a vector. (It can, but things go wrong.)}
}
\value{
A plot combining polygons and points.
   If \code{state.num} is not found within the dataset selected with
   \code{year}, no plot is created and an error is returned.
   If the combination of \code{state.num} and \code{year} is valid, but there
   are no accident data, the function returns NULL, but not made visible in
   the console, while an appropriate message (not error) is displayed.
}
\description{
This function attempts to output a polygon plot of a specific USA state with
points ("." by setting \code{pch} to 46 in the \code{\link{points}} function)
on the same plot indicating the location of all logged accidents within
a specified year. It automatically assigns \code{NA} values to off-planet
LAT/LONG combinations with the base R function \code{\link{is.na}}. The
values \code{xlim} and \code{ylim} of \code{link{map}} is bound by the most
outlying LAT/LONG points found within the data and subsequently the polygon
drawn is clipped beyond these bounds.
}
\examples{
\dontrun{
fars_map_state(1,2015)
fars_map_state(54,2015)
}

}
