% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read data from external source and output it}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string providing the name of the file containing
the wanted data.}
}
\value{
This function returns a tibble data frame as generated by first using
   the \code{\link{read_csv}} function contained in the \code{readr} package
   and then passing the result on to the \code{dplyr} package function
   \code{\link{tbl_df}}. If the provided filename does not exist, an
   appropriate error is shown.
}
\description{
This function returns a tibble data frame extracted from a CSV file or a CSV
file encoded in a .bz2 file. It works best used in conjunction with the
\code{make_filename} function to provide a correctly formatted input.
Note that it works within the current working directory unless a full path
is specified.
}
\examples{
fars_read("C:/Users/All/Documents/Testy/R/data/accident_2013.csv.bz2")
data <- fars_read(make_filename(2013))

}
