% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the dataset for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numbers or character strings representing years.}
}
\value{
This function returns a list (the same length as \code{years}) of
   tibbles. NULL is returned for each input year not corresponding to a valid
   dataset.
}
\description{
This function builds on the functionality of \code{make_filename} and
\code{fars_read} to read the data from multiple CSV files containing
the data for the years corresponding to the input years. The \code{dplyr}
functions \code{\link{mutate}} and \code{\link{select}}are used to tidy the
data and only return the 'MONTH' and 'year' variable collumns respectively.
The base R function\code{\link{tryCatch}}is used to catch errors created by
\code{\link{make_filename}} and show these errors and the end of running
without interrupting the execution of the base R function \code{\link{lapply}}.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c("2013","2014"))
data <- fars_read_years(2013)
}

}
