% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Returns ordered and summarized version of datasets}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numbers or character strings representing years.}
}
\value{
This function returns a tibble summarizing the data found for each
   input year requested. A warning is returned for each input year not
   corresponding to a valid dataset, while not affecting the output for valid
   years.
}
\description{
This function builds on the functionality of \code{fars_read_years} to read
the data from multiple CSV files containing
the data for the years corresponding to the input years. The \code{dplyr} and
\code{tidyr}
functions \code{\link{bind_rows}}, \code{\link{group_by}},
\code{\link{group_by}} and \code{\link{spread}} are used to tidy the
data and only return the 'MONTH' and 'year' variable columns respectively,
while ensuring that the 'MONTH' column is not duplicated when multiple years
are used as inputs.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_sumarize_years(c("2013","2014"))
data <- fars_summarize_years(2013)
}

}
