% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Create a full file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number that represent the year of the data file to be accessed.}
}
\value{
This function returns a character vector which is formatted to
   correspond to a filename of a specific year's dataset. The output is printed
   to the console.
}
\description{
This is a simple function that, creates a full filename compatible with the
US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System datasets. A year is taken as an input and inserted in the
file name. The input \code{year} is coerced as an integer with the R function
base function \code{\link{as.integer}}
}
\examples{
\dontrun{
make_filename(2013)
when<- 2013
make_filename(when)
make_filename("2013")
data <- make_filename(2013)
}

}
