% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testy.R
\name{testy}
\alias{testy}
\title{A function for testing your own function}
\usage{
testy(func)
}
\arguments{
\item{func}{A function in testing}
}
\description{
This function allows you to parse the default parameters in the function you are testing.
}
